requirements.txt
# Core dependencies
requests>=2.31.0
requests-cache>=1.1.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.12.0

# Geospatial
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0

# Date/time handling
python-dateutil>=2.8.2
pytz>=2024.1

# Testing
pytest>=7.4.0

# Type hints
typing-extensions>=4.9.0
